todo_backend/requirements.txt
asgiref==3.7.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.3
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
filelock==3.13.1
idna==3.9
mysql-connector-python==8.4.0
mysqlclient==2.2.5
pipenv==2024.0.2
platformdirs==4.2.0
pycparser==2.22
PyMySQL==1.1.1
setuptools==75.2.0
sqlparse==0.4.4
tk==0.1.0
tzdata==2023.4
urllib3==2.2.3
virtualenv==20.26.4
virtualenvwrapper-win==1.2.7
whitenoise==6.7.0
